Preserve dimension during subset
When the number of types with counts > 1  is < 2 (i.e. only one type with count > 1), the subsets resulted in a single column matrix, which was automatically simplified, dropping a dimension and resulting in the error `Error in array(STATS, dims[perm]) : 'dims' cannot be of length 0`. By specifying `drop = FALSE`, we preserve dimensionality and clear the error.